+++ src/physics/geom.h
@@ -1,13 +1,15 @@
 #ifndef _GEOM_H
 #define _GEOM_H
 
+#include <stdint.h>
+
 typedef struct Rect {
-    int x, y;
-    int w, h;
+    uint16_t x, y;
+    uint8_t w, h;
 } Rect;
 
 typedef struct Vec2 {
-    float x, y;
+    uint16_t x, y;
 } Vec2;
 
 #endif  // _GEOM_H
